+++ src/styles/hero.css
@@ -1,12 +1,15 @@
-.hero-container {
-    max-height: 70vh;
-    z-index: 1;
-    background-color: rgba(0,0,0, 0.5);
+.hero-img, .hero-img-inverse {
+    width: 70vw;
+    height: 100%;
+    max-height: 80vh;
 }
 
 .hero-img {
-    width: 50vw;
-    height: 100%;
-    -webkit-clip-path: polygon(15% 0, 100% 0%, 100% 100%, 0% 100%);
-    clip-path: polygon(15% 0, 100% 0%, 100% 100%, 0% 100%);
+    -webkit-clip-path: polygon(20% 0, 100% 0%, 100% 100%, 0% 100%);
+    clip-path: polygon(20% 0, 100% 0%, 100% 100%, 0% 100%);
+}
+
+.hero-img-inverse {
+    -webkit-clip-path: polygon(0 0, 100% 0, 80% 100%, 0% 100%);
+    clip-path: polygon(0 0, 100% 0, 80% 100%, 0% 100%);
 }
