+++ src/styles/heroContainer.css
@@ -1,0 +1,8 @@
+.vanta-canvas {
+    z-index: -1 !important;
+}
+
+.hero-container {
+    z-index: 1;
+    background-color: rgba(0,0,0, 0.5);
+}
